refactor(): create subfolders
- cards
- table
