+++ src/main/kotlin/blackjack/model/Results.kt
@@ -1,5 +1,9 @@
 package blackjack.model
 
+import blackjack.model.table.Dealer
+import blackjack.model.table.Participant
+import blackjack.model.table.Players
+
 class Results(private val dealer: Dealer, private val players: Players){
     fun calculate(): Pair<Double, Map<Participant, Double>>{
         var dealerResult = 0.0
