+++ src/main/kotlin/blackjack/model/states/Hit.kt
@@ -1,7 +1,7 @@
 package blackjack.model.states
 
-import blackjack.model.Card
-import blackjack.model.HandCards
+import blackjack.model.cards.Card
+import blackjack.model.cards.HandCards
 
 class Hit(override val hand: HandCards) : Running {
     init {
