+++ src/main/kotlin/blackjack/model/table/Player.kt
@@ -1,12 +1,14 @@
-package blackjack.model
+package blackjack.model.table
 
+import blackjack.model.Bet
+import blackjack.view.ErrorMessage
 import blackjack.model.states.Init
 import blackjack.model.states.State
 
 class Player(
     override val name: String,
     override var state: State = Init(),
-    var bet: Bet = Bet.UNPLACED,
+    var bet: Bet = Bet.Companion.UNPLACED,
 ) : Participant() {
     init {
         require(name.isNotBlank()) { ErrorMessage.NAME_BLANK.toString() }
