+++ src/main/kotlin/blackjack/view/ErrorMessage.kt
@@ -1,10 +1,11 @@
-package blackjack.model
+package blackjack.view
 
 import blackjack.controller.Controller
+import blackjack.model.table.Players
 
 enum class ErrorMessage(val message: String) {
-    MAX_TRIES("Maximum tries of ${Controller.MAX_TRIES} reached."),
-    PLAYER_RANGE("Only ${Players.MIN} - ${Players.MAX} players accepted."),
+    MAX_TRIES("Maximum tries of ${Controller.Companion.MAX_TRIES} reached."),
+    PLAYER_RANGE("Only ${Players.Companion.MIN} - ${Players.Companion.MAX} players accepted."),
     EMPTY_INPUT("Input is empty."),
     BETTING_AMOUNT("Betting amount has to be a positive integer"),
     NAME_BLANK("Name cannot be blank."),
