+++ src/test/kotlin/blackjack/HitTest.kt
@@ -2,7 +2,7 @@ package blackjack
 
 import blackjack.model.states.BlackJack
 import blackjack.model.states.Bust
-import blackjack.model.HandCards
+import blackjack.model.cards.HandCards
 import blackjack.model.states.State
 import blackjack.model.states.Hit
 import blackjack.model.states.Stay
