+++ src/test/kotlin/blackjack/InitTest.kt
@@ -4,7 +4,7 @@ import blackjack.model.states.State
 import blackjack.model.states.Init
 import blackjack.model.states.Hit
 import blackjack.model.states.BlackJack
-import blackjack.model.HandCards
+import blackjack.model.cards.HandCards
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
